fix: custom hook useUsers
