+++ src/redux/slices/users/hooks/useUsers.js
@@ -1,0 +1,18 @@
+import React, { useState, useEffect } from "react";
+import { useDispatch, useSelector } from "react-redux";
+import { getUsersList } from "..";
+
+export const useUsers = () => {
+  const { usersData } = useSelector((state) => state.users);
+
+  const [pageState, setPageState] = useState(
+    usersData.page ? usersData.page : 1
+  );
+  const dispatch = useDispatch();
+
+  useEffect(() => {
+    dispatch(getUsersList(pageState));
+  }, [dispatch, pageState]);
+
+  return { usersData, pageState, setPageState };
+};
